requirements.txt
streamlit
pillow
numpy
scikit-image
requests
streamlit-image-comparison